bear-detection-backend2-0703/requirements.txt
Flask
Flask-Cors
requests
Pillow
moviepy
opencv-python
